Remove all the WithBlinding function in favor of an optional blinder.
Every operations with private key (signing and decrypt) now take an
optional blinder. reduce the number of API calls substantially.

Blinder should be set when timing is important to not disclose
any information; If any doubt, the blinder should be set,
as the cost is negligable, and you can compute blinders in advance.